requirements.txt
aws-cdk-lib==2.60.0
boto3==1.24.43
requests==2.28.0
requests-aws4auth==1.1.2
requests-oauthlib==1.3.1

